TemporaryFile: fix move assignment
Fix the move assignment operator of `TemporaryFile`, which would behave
incorrectly if the object was moved into itself.

+++ include/revng/Support/TemporaryFile.h
@@ -41,6 +41,9 @@ public:
 
   TemporaryFile(TemporaryFile &&Other) { *this = std::move(Other); }
   TemporaryFile &operator=(TemporaryFile &&Other) {
+    if (this == &Other)
+      return *this;
+
     if (not Path.empty()) {
       revng::cantFail(llvm::sys::fs::remove(Path));
       llvm::sys::DontRemoveFileOnSignal(Path);
